fix: finally fix up file rendering
