+++ packages/app/app/components/Message/Model/FileRenderer/Audio.vue
@@ -1,11 +1,11 @@
 <script setup lang="ts">
 import type { FileRendererComponentProps } from "@/models/message/file/FileRendererComponentProps";
 
-const { isPreview, url } = defineProps<FileRendererComponentProps>();
+const { url } = defineProps<FileRendererComponentProps>();
 </script>
 
 <template>
-  <StyledCard p-2 flex items-center justify-center :class="isPreview ? 'size-full' : undefined">
+  <StyledCard p-2 flex items-center justify-center size-full>
     <audio :src="url" controls w-full />
   </StyledCard>
 </template>
