+++ packages/app/app/components/Message/Model/FileRenderer/Index.vue
@@ -4,7 +4,6 @@ import type { FileRendererComponentProps } from "@/models/message/file/FileRende
 import { TypeRendererMap } from "@/models/message/file/TypeRendererMap";
 import { getLanguage } from "@/services/codemirror/getLanguage";
 import { getInferredMimetype } from "@/services/file/getInferredMimetype";
-import { PREVIEW_MAX_HEIGHT } from "@/services/message/file/constants";
 import { takeOne } from "@esposter/shared";
 
 const props = defineProps<FileRendererComponentProps>();
@@ -21,14 +20,5 @@ const renderer = computed<Component>(() => {
 </script>
 
 <template>
-  <v-skeleton-loader
-    v-if="!props.url"
-    size-full
-    :height="props.isPreview ? PREVIEW_MAX_HEIGHT : '16rem'"
-    type="image"
-  />
-  <div v-else-if="props.isPreview" bg-surface flex size-full min-h-0 items-center justify-center overflow-hidden>
-    <component :is="renderer" :language :="props" />
-  </div>
-  <component :is="renderer" v-else :language :="props" />
+  <component :is="renderer" :language :="props" />
 </template>
