+++ packages/app/app/components/Message/Model/Message/File/Container.vue
@@ -13,8 +13,8 @@ const columnLayout = computed(() => getColumnLayout(message.files.length));
 </script>
 
 <template>
-  <v-row m--0.25>
-    <v-col v-for="(file, index) of message.files" :key="file.id" p-0.25 :cols="columnLayout[index]">
+  <v-row no-gutters>
+    <v-col v-for="(file, index) of message.files" :key="file.id" p-0.5 :cols="columnLayout[index]">
       <MessageModelMessageFile :column-layout :file :index :is-preview :message />
     </v-col>
   </v-row>
